+++ .gitignore
@@ -11,4 +11,4 @@ bundled.js
 .bundled.js
 cycledash/static/lib/
 coverage.html
-/cycledash/static/dalliance/
+
