+++ cycledash/projects.py
@@ -1,17 +1,19 @@
-from flask import redirect, jsonify, url_for
+"""Defines the API for Projects."""
+from flask import request, redirect, jsonify, url_for, render_template
 
 from sqlalchemy import exc, select, func, desc
 import voluptuous
 
-from common.helpers import tables
+from common.helpers import tables, find
 from cycledash import db
+import cycledash.tasks
 import cycledash.validations
+import cycledash.comments
 from cycledash.helpers import (prepare_request_data, error_response,
                                request_wants_json, get_where, get_id_where)
 
 
 def get_project(project_id):
-    """Return project, or None is no matching project is found."""
     with tables(db, 'projects') as (con, projects):
         q = select(projects.c).where(projects.c.id == project_id)
         project = con.execute(q).fetchone()
@@ -29,15 +31,13 @@ def get_projects():
         return jsonify({'projects': [dict(r) for r in con.execute(q).fetchall()]})
 
 
-def create_project(request):
+def create_project():
     """Create a project, return the project."""
     try:
-        data = cycledash.validations.CreateProjectSchema(
+        data = cycledash.validations.CreateProject(
             prepare_request_data(request))
     except voluptuous.MultipleInvalid as e:
         errors = [str(err) for err in e.errors]
-        if len(errors) == 1:
-            errors = errors[0]
         return error_response('Project validation', errors)
     try:
         with tables(db, 'projects') as (con, projects):
@@ -47,20 +47,18 @@ def create_project(request):
         return error_response('Could not create project {}'.format(data),
                               str(e))
     if request_wants_json():
-        return jsonify(project), 201
+        return jsonify(project), 201  # HTTP 201 Created
     elif 'text/html' in request.accept_mimetypes:
         return redirect(url_for('list_runs'))
 
 
-def update_project(project_id, request):
+def update_project(project_id):
     """Update the project, return the updated project."""
     try:
-        data = cycledash.validations.UpdateProjectSchema(
+        data = cycledash.validations.UpdateProject(
             prepare_request_data(request))
     except voluptuous.MultipleInvalid as e:
         errors = [str(err) for err in e.errors]
-        if len(errors) == 1:
-            errors = errors[0]
         return error_response('Project validation', errors)
     try:
         with tables(db, 'projects') as (con, projects):
@@ -104,3 +102,74 @@ def set_and_verify_project_id_on(data):
         project_id = data['project_id']
         if get_where('projects', db, id=project_id) is None:
             raise voluptuous.Invalid('no project with id {}'.format(project_id))
+
+
+def _get_projects_tree():
+    """Return a list of all projects, with their respective vcfs and bams.
+
+    { "projects": [
+      { "name": "Project Name",
+        "notes": "Some test notes"
+        "vcfs": [...]
+        "bams": [
+          { "name": "a dataset", ... }, ...
+      ]}, ...
+    ]}
+    """
+    with tables(db, 'vcfs', 'user_comments', 'bams', 'projects') as \
+         (con, vcfs, user_comments, bams, projects):
+        joined = (vcfs
+            .outerjoin(user_comments, vcfs.c.id == user_comments.c.vcf_id))
+        num_comments = func.count(user_comments.c.vcf_id).label('num_comments')
+        q = (select(vcfs.c + [num_comments])
+            .select_from(joined)
+            .group_by(vcfs.c.id)
+            .order_by(desc(vcfs.c.id)))
+        vcfs = [dict(v) for v in con.execute(q).fetchall()]
+
+        q = select(bams.c)
+        bams = [dict(b) for b in con.execute(q).fetchall()]
+
+        q = select(projects.c)
+        projects = [dict(b) for b in con.execute(q).fetchall()]
+
+        for vcf in vcfs:
+            normal_bam_id = vcf.get('normal_bam_id')
+            tumor_bam_id = vcf.get('tumor_bam_id')
+            project_id = vcf.get('project_id')
+
+            vcf['project'] = dict(find(projects,
+                                       lambda x: x.get('id') == project_id) or {})
+            vcf['tumor_bam'] = dict(find(bams,
+                                         lambda x: x.get('id') == tumor_bam_id) or {})
+            vcf['normal_bam'] = dict(find(bams,
+                                          lambda x: x.get('id') == normal_bam_id) or {})
+        _join_task_states(vcfs)
+
+        for project in projects:
+            project_id = project['id']
+            project_bams = [bam for bam in bams
+                            if bam.get('project_id') == project_id]
+            project_vcfs = [vcf for vcf in vcfs
+                            if vcf.get('project_id') == project_id]
+            project['bams'] = project_bams
+            project['vcfs'] = project_vcfs
+        return projects
+
+
+def _join_task_states(vcfs):
+    """Add a task_states field to each VCF in a list of VCFs."""
+    ts = cycledash.tasks.all_non_success_tasks()
+
+    for vcf in vcfs:
+        vcf['task_states'] = ts.get(vcf['id'], [])
+
+
+def get_projects_tree():
+    if request_wants_json():
+        vcfs = _get_projects_tree()
+        return jsonify({'runs': vcfs})
+    elif 'text/html' in request.accept_mimetypes:
+        vcfs = _get_projects_tree()
+        comments = cycledash.comments.get_last_comments()
+        return render_template('runs.html', last_comments=comments, runs=vcfs)
