+++ cycledash/templates/about.html
@@ -56,7 +56,7 @@
         <dt>variantCallerName</dt>
         <dd>
           The name of the variant caller which produced this VCF. This
-          should remain constrant between VCFs with the same caller in
+          should remain constant between VCFs with the same caller in
           order to compare runs to one another.
         </dd>
       </dl>
@@ -108,12 +108,6 @@
         <dd>The resected tissue types.</dd>
         <dt>resection_date</dt>
         <dd>The resected date, in YYYY-MM-DD form.</dd>
-        <dt>library</dt>
-        <dd>The library used for sequencing.</dd>
-        <dt>sequencing_platform</dt>
-        <dd>e.g. Illumina, PacBio, etc.</dd>
-        <dt>primary_cancer_site</dt>
-        <dd>true or false</dd>
       </dl>
     </section>
   </section>
